+++ tests/v1/mod.rs
@@ -3,22 +3,24 @@
 // Copyright: 2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+pub mod invites;
 pub mod members;
 pub mod server;
 pub mod workspace;
 
 use cucumber::{given, then, when};
-use entity::model::{MemberRole, JID};
+use entity::model::{self, MemberRole};
 use entity::{member, server_config};
 use prose_pod_api::error::Error;
 use prose_pod_api::guards::JWTService;
-use prose_pod_api::v1::InitRequest;
+use prose_pod_api::v1::{AdminAccountInit, InitRequest};
 use rocket::http::{ContentType, Status};
 use rocket::local::asynchronous::{Client, LocalResponse};
 use serde_json::json;
 use service::sea_orm::{ActiveModelTrait, Set};
 use service::Mutation;
 
+use crate::cucumber_parameters::JID;
 use crate::TestWorld;
 
 pub const DEFAULT_WORKSPACE_NAME: &'static str = "Prose";
@@ -27,7 +29,7 @@ pub const DEFAULT_WORKSPACE_NAME: &'static str = "Prose";
 async fn given_admin(world: &mut TestWorld, name: String) -> Result<(), Error> {
     let db = world.db();
 
-    let jid = JID::new(name.to_lowercase().replace(" ", "-"), "test.org");
+    let jid = model::JID::new(name.to_lowercase().replace(" ", "-"), "test.org");
 
     let model = member::ActiveModel {
         id: Set(jid.to_string()),
@@ -48,7 +50,7 @@ async fn given_admin(world: &mut TestWorld, name: String) -> Result<(), Error> {
 async fn given_not_admin(world: &mut TestWorld, name: String) -> Result<(), Error> {
     let db = world.db();
 
-    let jid = JID::new(name.to_lowercase().replace(" ", "-"), "test.org");
+    let jid = model::JID::new(name.to_lowercase().replace(" ", "-"), "test.org");
 
     let model = member::ActiveModel {
         id: Set(jid.to_string()),
@@ -137,13 +139,22 @@ async fn then_error_workspace_already_initialized(world: &mut TestWorld) {
 
 // INIT
 
-async fn init_workspace<'a>(client: &'a Client, name: &str) -> LocalResponse<'a> {
+async fn init_workspace<'a>(
+    client: &'a Client,
+    admin_jid: &model::JID,
+    name: &str,
+) -> LocalResponse<'a> {
     client
         .post("/v1/init")
         .header(ContentType::JSON)
         .body(
             json!(InitRequest {
                 workspace_name: name.to_string(),
+                admin: AdminAccountInit {
+                    jid: admin_jid.clone(),
+                    password: "password".to_string(),
+                    nickname: admin_jid.node.replace(".", " "),
+                }
             })
             .to_string(),
         )
@@ -151,9 +162,9 @@ async fn init_workspace<'a>(client: &'a Client, name: &str) -> LocalResponse<'a>
         .await
 }
 
-#[when(expr = "a user initializes a workspace named {string}")]
-async fn when_workspace_init(world: &mut TestWorld, name: String) {
-    let res = init_workspace(&world.client, &name).await;
+#[when(expr = "<{jid}> initializes a workspace named {string}")]
+async fn when_workspace_init(world: &mut TestWorld, jid: JID, name: String) {
+    let res = init_workspace(&world.client, &jid, &name).await;
     world.result = Some(res.into());
 }
 
@@ -164,6 +175,11 @@ async fn test_init_workspace() -> Result<(), Box<dyn Error>> {
     let workspace_name = DEFAULT_WORKSPACE_NAME;
     let body = serde_json::to_string(&InitRequest {
         workspace_name: workspace_name.to_string(),
+        admin: AdminAccountInit {
+            jid: "test.admin@prose.org",
+            password: "password".to_string(),
+            nickname: "Test admin".to_string(),
+        },
     })?;
 
     let response = client.post("/v1/init").body(&body).dispatch().await;
